// frontend-template/src/api/enums/http-request-method-enum.ts
/**
 * Fast.IaaS.HttpRequestMethodEnum Http请求方式枚举
 * @export
 * @enum {string}
 */
export enum HttpRequestMethodEnum {
    /**
     * Get请求
     */
    Get = 1,
    /**
     * Post请求
     */
    Post = 2,
    /**
     * Put请求
     */
    Put = 3,
    /**
     * Delete请求
     */
    Delete = 4,
    /**
     * Patch请求
     */
    Patch = 5,
    /**
     * Head请求
     */
    Head = 6,
    /**
     * Options请求
     */
    Options = 7,
    /**
     * Connect请求
     */
    Connect = 8,
    /**
     * Trace请求
     */
    Trace = 9,
}